requirements.txt
# Add your Python dependencies here, one per line.
flask
requests
python-dotenv
psycopg2-binary
gunicorn
